+++ CHANGELOG.md
@@ -50,6 +50,7 @@ Legacy code under hpOneView which was marked as deprecated has been removed. Thi
 - Server profile template
 - Storage pool
 - Storage system
+- Storage volume
 - Storage volume attachment
 - Switch
 - Switch type